server/requirements.txt
fastapi
uvicorn
pydantic
pandas
dvc[s3,ssh]==3.51.1
python-multipart
sqlalchemy
asyncpg
python-dotenv
psycopg[binary]
httpx
jsonpath_ng
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
itsdangerous>=2.1.0
